# Skip defined instruction in NODE_OP_ASGN_OR with ivar